Add support for game-specific emotes
This is hacky as fuck. A rewrite is coming, I promise.

The way you use this is by tossing a JSON object into the customEmotes field in the discord_channels config store that is roughly this: keys are channel IDs, which map to objects whose keys are emoji IDs or the respective raw unicode characters, and those keys map to role IDs to toggle.